+++ lwt/tls_lwt.mli
@@ -25,7 +25,6 @@ module Unix : sig
   val write_bytes : t -> Lwt_bytes.t -> int -> int -> unit Lwt.t
 
   val epoch  : t -> [ `Ok of Tls.Engine.epoch_data | `Error ]
-
 end
 
 type ic = Lwt_io.input_channel
@@ -36,7 +35,7 @@ val accept_ext :
   ((ic * oc) * Lwt_unix.sockaddr) Lwt.t
 
 val accept :
-  ?trace:tracer -> X509_lwt.priv -> Lwt_unix.file_descr ->
+  ?trace:tracer -> Tls.Config.own_cert -> Lwt_unix.file_descr ->
   ((ic * oc) * Lwt_unix.sockaddr) Lwt.t
 
 val connect_ext :